+++ src/main/java/de/codescape/jira/plugins/scrumpoker/upgrade/Upgrade02DefaultProjectActivation.java
@@ -2,6 +2,7 @@ package de.codescape.jira.plugins.scrumpoker.upgrade;
 
 import com.atlassian.plugin.spring.scanner.annotation.export.ExportAsService;
 import com.atlassian.sal.api.upgrade.PluginUpgradeTask;
+import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSettingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -34,7 +35,9 @@ public class Upgrade02DefaultProjectActivation extends AbstractUpgradeTask {
 
     @Override
     protected void performUpgrade() {
-        scrumPokerSettingService.persistDefaultProjectActivation(true);
+        GlobalSettings globalSettings = scrumPokerSettingService.load();
+        globalSettings.setDefaultProjectActivation(true);
+        scrumPokerSettingService.persist(globalSettings);
     }
 
 }